Added some old functions from last app to get the correct images and cardinal wind information.

+++ app/src/main/java/com/scoutzknifez/weatherappv2/DataStructures/WeatherParent.java
@@ -14,7 +14,7 @@ public abstract class WeatherParent {
     private double precipitationProbability;
     private double humidity;
     private int windSpeed;
-    private int windBearing;
+    private double windBearing;
 
     @Override
     public String toString() {
